code shuffling pig farm working
